hack/client/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jedi==0.18.0
parso==0.8.1
pytz==2021.1
requests==2.25.1
rfc5424-logging-handler==1.4.3
six==1.15.0
syslog-rfc5424-formatter==1.2.2
tzlocal==2.1
urllib3==1.26.3
websocket-client==0.57.0
